remade enemy move
simplified enemy move. remove coordinates abstraction level.

+++ Assets/Scripts/AttackStats.cs
@@ -6,7 +6,7 @@ public struct AttackStats
 {
     public int MinDamage { get; set; }
     public int MaxDamage { get; set; }
-    public float ArmorPiercing { get; set; }
+    public int ArmorPiercing { get; set; }
     public float SlowStrength { get; set; }
     public int Heal { get; set; }
     public float Range { get; set; }
